requirements.txt
django-cors-headers==4.2.0
djangorestframework==3.14.0
Pillow==10.0.1
drf-yasg==1.21.7
